requirements.txt
# AutoGen v0.7.5+ packages
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext[openai,azure]==0.7.5

# Azure SDKs
azure-identity==1.19.0

# MongoDB for Cosmos DB vCore
pymongo==4.6.1

# PDF Processing
PyPDF2==3.0.1

# Environment Management
python-dotenv==1.0.1

# Optional: Better PDF parsing
# pypdf==5.1.0
# pdfplumber==0.11.4
